Remove dependencies & clean up
