+++ src/main/java/com/imjustdoom/pluginsite/config/security/SecurityConfig.java
@@ -1,50 +1,37 @@
 package com.imjustdoom.pluginsite.config.security;
 
 import lombok.AllArgsConstructor;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
 import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.password.PasswordEncoder;
 
 @Configuration
 @EnableWebSecurity
 @AllArgsConstructor
 public class SecurityConfig extends WebSecurityConfigurerAdapter {
-
-    private final UserDetailsService userDetailsService;
-    private final PasswordEncoder passwordEncoder;
-
-    @Bean
-    public DaoAuthenticationProvider authProvider() {
-        DaoAuthenticationProvider authProvider = new DaoAuthenticationProvider();
-        authProvider.setUserDetailsService(this.userDetailsService);
-        authProvider.setPasswordEncoder(this.passwordEncoder);
-        return authProvider;
-    }
+    private final DaoAuthenticationProvider authProvider;
 
     @Override
     protected void configure(AuthenticationManagerBuilder auth) {
-        auth.authenticationProvider(authProvider());
+        auth.authenticationProvider(this.authProvider);
     }
 
     @Override
     public void configure(HttpSecurity http) throws Exception {
         http
-                .csrf().disable()
+            .csrf().disable()
 
-                .authorizeRequests()
-                .antMatchers("/admin", "/admin/roles").hasRole("ADMIN")
-                .antMatchers("/resources/create", "/account/details").authenticated()
-                .antMatchers("/signup", "/login").not().authenticated()
+            .authorizeRequests()
+            .antMatchers("/admin", "/admin/roles").hasRole("ADMIN")
+            .antMatchers("/resources/create", "/account/details").authenticated()
+            .antMatchers("/register", "/login").not().authenticated()
 
-                .anyRequest().permitAll()
+            .anyRequest().permitAll()
 
-                .and()
-                .formLogin().loginPage("/login");
+            .and()
+            .formLogin().loginProcessingUrl("/login");
     }
 }
