--- src/main/java/com/imjustdoom/pluginsite/util/StringUtil.java
@@ -1,15 +1,0 @@
-package com.imjustdoom.pluginsite.util;
-
-import org.commonmark.node.Node;
-import org.commonmark.parser.Parser;
-import org.commonmark.renderer.html.HtmlRenderer;
-
-public class StringUtil {
-
-    public static String markdownToHtml(String markdown) {
-        Parser parser = Parser.builder().build();
-        Node document = parser.parse(markdown);
-        HtmlRenderer renderer = HtmlRenderer.builder().build();
-        return renderer.render(document);
-    }
-}
